Modularize the install command to use board-provided installers. This paves the way to adding more install targets (like real chips or VCS or something).
