+++ boards/default/installers/firesim/__init__.py
@@ -1,0 +1,1 @@
+from .firesim import *
