+++ boards/default/installers/firesim/dummy.rootfs
