+++ boards/firechip/installers
@@ -1,0 +1,1 @@
+../default/installers/
